# internal/api/v1/docs/authAPI_swagger.yaml
basePath: /v1
definitions:
  models.ConfigResponse:
    properties:
      default_next_url:
        type: string
      device_enabled:
        type: boolean
      endpoint:
        type: string
      github_enabled:
        type: boolean
      google_enabled:
        type: boolean
      magic_enabled:
        type: boolean
    type: object
...
  models.ServiceKeyLoginResponse:
    properties:
      organization:
        type: string
      resource_id:
        type: string
      resource_type:
        type: string
      service_key_id:
        type: string
      service_session_id:
        type: string
      service_session_secret:
        type: string
      user_id:
        type: string
    type: object
  models.UserInfoResponse:
    properties:
      email:
        type: string
      organization:
        type: string
      session:
        enum:
        - session
        - api
        - service
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: admin@loopholelabs.io
    name: API Support
  description: Auth API, V1
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://loopholelabs.io/privacy
  title: Auth API V1
  version: "1.0"
paths:
  /config:
    get:
      consumes:
      - application/json
      description: Config gets the public configuration of the API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ConfigResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Config gets the public configuration of the API
      tags:
      - config
...
  /logout:
    post:
      consumes:
      - application/json
      description: Logout logs out a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Logout logs out a user
      tags:
      - logout
  /magic/callback:
    get:
      consumes:
      - application/json
      description: MagicCallback validates the magic link and logs in the user
      parameters:
      - description: magic link token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "307":
          description: Temporary Redirect
          headers:
            Location:
              description: Redirects to Next URL
              type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: MagicCallback validates the magic link and logs in the user
      tags:
      - magic
      - callback
  /magic/flow:
    post:
      consumes:
      - application/json
      description: MagicFlow starts the magic link flow
      parameters:
      - description: email address
        in: query
        name: email
        required: true
        type: string
      - description: Next Redirect URL
        in: query
        name: next
        type: string
      - description: Organization
        in: query
        name: organization
        type: string
      - description: Device Flow Identifier
        in: query
        name: identifier
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: MagicFlow starts the magic link flow
      tags:
      - device
      - login
  /servicekey/login:
    post:
      consumes:
      - application/json
      description: ServiceKeyLogin logs in a user with their Service Key
      parameters:
      - description: Service Key
        in: query
        name: servicekey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ServiceKeyLoginResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: ServiceKeyLogin logs in a user with their Service Key
      tags:
      - servicekey
      - login
  /userinfo:
    post:
      consumes:
      - application/json
      description: UserInfo checks if a user is logged in and returns their info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserInfoResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: UserInfo checks if a user is logged in and returns their info
      tags:
      - login
schemes:
- https
swagger: "2.0"